Update JadcupPLC.st
add functionality of the second machine in factory IO. the current code has a block of testing code due to missing functionality in the middleware. currently jobs queued to machine 1 are also queued to m2. additionally M1 and M2 running code is being auto set to true.

remove this if you want to test middleware code